Use a gentoo mirrors data file
This commit changes the way mirrors are obtained. Previously the gentoo
mirrors HTML page was parsed for mirrors, but this is way too slow to do
per command. Instead the mirrors are saved in a JSON format in the data
folder. These will be the default mirrors. Fresh mirrors will be pulled
down and stored when the script is first installed.